Integ-tests: move `render_jinja_template` to common file
