+++ tests/integration-tests/tests/common/osu_common.py
@@ -10,13 +10,12 @@
 # This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, express or implied.
 # See the License for the specific language governing permissions and limitations under the License.
 import logging
-import os as os_lib
 import pathlib
 import re
 from shutil import copyfile
 
 from constants import OSU_BENCHMARK_VERSION
-from jinja2 import Environment, FileSystemLoader
+from utils import render_jinja_template
 
 from tests.common.utils import fetch_instance_slots
